Feature/picos integration new task flow

See merge request ompss-at-fpga/ompss-manager-src!24

+++ generate_IPs.py
@@ -32,19 +32,21 @@ from distutils import spawn
 import xml.etree.cElementTree as cET
 
 SOM_MAJOR_VERSION = 2
-SOM_MINOR_VERSION = 0
-SOM_PREVIOUS_MAJOR_VERSION = 1
-SOM_PREVIOUS_MINOR_VERSION = 6
+SOM_MINOR_VERSION = 1
 
-POM_MAJOR_VERSION = 0
+SOM_PREVIOUS_MAJOR_VERSION = 2
+SOM_PREVIOUS_MINOR_VERSION = 0
+
+POM_MAJOR_VERSION = 1
 POM_MINOR_VERSION = 0
+
 POM_PREVIOUS_MAJOR_VERSION = 0
 POM_PREVIOUS_MINOR_VERSION = 0
 
 class Logger(object):
     def __init__(self):
         self.terminal = sys.stdout
-        self.log = open('./som_IP.log', 'w+')
+        self.log = open('generate_IPs.log', 'w+')
         self.subprocess = subprocess.PIPE if args.verbose else self.log
         self.re_color = re.compile(r'\033\[[0,1][0-9,;]*m')
 
@@ -66,7 +68,7 @@ class Color:
 
 class Messages:
     def error(self, msg):
-        print(Color.RED + msg + '. Check som_IP.log for more information' + Color.END)
+        print(Color.RED + msg + '. Check generate_IPs.log for more information' + Color.END)
         sys.exit(1)
 
     def info(self, msg):
@@ -90,6 +92,10 @@ class ArgParser:
         self.parser.add_argument('-c', '--clock', help='FPGA clock frequency in MHz\n(def: \'100\')', type=int, default='100')
         self.parser.add_argument('-v', '--verbose', help='prints Vivado messages', action='store_true', default=False)
         self.parser.add_argument('--skip_hls', help='skips the cleanup and HLS step', action='store_true', default=False)
+        self.parser.add_argument('--skip_board_check', help='skips the board part check', action='store_true', default=False)
+        self.parser.add_argument('--skip_cutoff_gen', help='skips generation of the CutoffManager IP', action='store_true', default=False)
+        self.parser.add_argument('--skip_pom_gen', help='skips generation of the PicosOmpSsManager IP', action='store_true', default=False)
+        self.parser.add_argument('--skip_som_gen', help='skips generation of the SmartOmpSsManager IP', action='store_true', default=False)
 
     def parse_args(self):
         return self.parser.parse_args()
@@ -107,8 +113,55 @@ def compute_resource_utilization(acc_path, extended=False):
 
         used_resources[True][resource.tag] = int(resource.text) + (int(used_resources[extended][resource.tag]) if resource.tag in used_resources[extended] else 0)
 
+def generate_cutoff_IP():
+    msg.info('Generating CutoffManager IP')
 
-def generate_IP():
+    prj_path = './cutoff_IP/Vivado/CutoffManager'
+
+    p = subprocess.Popen('vivado -nojournal -nolog -notrace -mode batch -source '
+                         + os.getcwd() + '/scripts/cutoff_ip_packager.tcl -tclargs '
+                         + 'CutoffManager '
+                         + args.board_part + ' ' + os.getcwd() + ' '
+                         + os.path.abspath(os.getcwd() + '/cutoff_IP'), cwd=prj_path,
+                         stdout=sys.stdout.subprocess,
+                         stderr=sys.stdout.subprocess, shell=True)
+
+    if args.verbose:
+        for line in iter(p.stdout.readline, b''):
+            sys.stdout.write(line.decode('utf-8'))
+
+    retval = p.wait()
+    if retval:
+        msg.error('Generation of CutoffManager IP failed')
+    else:
+        msg.success('Finished generation of CutoffManager IP')
+
+def generate_POM_IP():
+    msg.info('Generating PicosOmpSsManager IP')
+
+    prj_path = './pom_IP/Vivado/PicosOmpSsManager'
+
+    p = subprocess.Popen('vivado -nojournal -nolog -notrace -mode batch -source '
+                         + os.getcwd() + '/scripts/pom_ip_packager.tcl -tclargs '
+                         + 'PicosOmpSsManager '
+                         + str(POM_MAJOR_VERSION) + '.' + str(POM_MINOR_VERSION) + ' '
+                         + str(POM_PREVIOUS_MAJOR_VERSION) + '.' + str(POM_PREVIOUS_MINOR_VERSION) + ' '
+                         + args.board_part + ' ' + os.getcwd() + ' '
+                         + os.path.abspath(os.getcwd() + '/pom_IP'), cwd=prj_path,
+                         stdout=sys.stdout.subprocess,
+                         stderr=sys.stdout.subprocess, shell=True)
+
+    if args.verbose:
+        for line in iter(p.stdout.readline, b''):
+            sys.stdout.write(line.decode('utf-8'))
+
+    retval = p.wait()
+    if retval:
+        msg.error('Generation of PicosOmpSsManager IP failed')
+    else:
+        msg.success('Finished generation of PicosOmpSsManager IP')
+
+def generate_SOM_IP():
     msg.info('Generating SmartOmpSsManager IP')
 
     prj_path = './som_IP/Vivado/SmartOmpSsManager'
@@ -116,8 +169,8 @@ def generate_IP():
     p = subprocess.Popen('vivado -nojournal -nolog -notrace -mode batch -source '
                          + os.getcwd() + '/scripts/som_ip_packager.tcl -tclargs '
                          + 'SmartOmpSsManager '
-                         + str(MAJOR_VERSION) + '.' + str(MINOR_VERSION) + ' '
-                         + str(PREVIOUS_MAJOR_VERSION) + '.' + str(PREVIOUS_MINOR_VERSION) + ' '
+                         + str(SOM_MAJOR_VERSION) + '.' + str(SOM_MINOR_VERSION) + ' '
+                         + str(SOM_PREVIOUS_MAJOR_VERSION) + '.' + str(SOM_PREVIOUS_MINOR_VERSION) + ' '
                          + args.board_part + ' ' + os.getcwd() + ' '
                          + os.path.abspath(os.getcwd() + '/som_IP'), cwd=prj_path,
                          stdout=sys.stdout.subprocess,
@@ -133,15 +186,14 @@ def generate_IP():
     else:
         msg.success('Finished generation of SmartOmpSsManager IP')
 
-
 def synthesize_hls(file_, includes, extended=False):
     acc_file = os.path.basename(file_)
     acc_name = os.path.splitext(acc_file)[0].replace('\.*', '')
 
     if extended:
-        dst_path = './som_IP/Vivado_HLS/extended/'
+        dst_path = './Vivado_HLS/extended/'
     else:
-        dst_path = './som_IP/Vivado_HLS/'
+        dst_path = './Vivado_HLS/'
 
     os.makedirs(dst_path + acc_name)
     shutil.copy2(file_, dst_path + acc_name + '/' + acc_file)
@@ -187,45 +239,74 @@ args = parser.parse_args()
 sys.stdout = Logger()
 
 if spawn.find_executable('vivado_hls') and spawn.find_executable('vivado'):
-    msg.info('Checking if your current version of Vivado supports the selected board part')
-    os.system('echo "if {[llength [get_parts ' + args.board_part + ']] == 0} {exit 1}" > ./board_part_check.tcl')
-    p = subprocess.Popen('vivado -nojournal -nolog -mode batch -source ./board_part_check.tcl', shell=True, stdout=open(os.devnull, 'w'))
-    retval = p.wait()
-    os.system('rm ./board_part_check.tcl')
-    if (int(retval) == 1):
-        msg.error('Your current version of Vivado does not support part ' + args.board_part)
+    if not args.skip_board_check:
+        msg.info('Checking if your current version of Vivado supports the selected board part')
+        os.system('echo "if {[llength [get_parts ' + args.board_part + ']] == 0} {exit 1}" > ./board_part_check.tcl')
+        p = subprocess.Popen('vivado -nojournal -nolog -mode batch -source ./board_part_check.tcl', shell=True, stdout=open(os.devnull, 'w'))
+        retval = p.wait()
+        os.system('rm ./board_part_check.tcl')
+        if (int(retval) == 1):
+            msg.error('Your current version of Vivado does not support part ' + args.board_part)
 
-    msg.success('Success')
+        msg.success('Success')
 
 else:
     msg.error('vivado_hls or vivado not found. Please set PATH correctly')
 
-if os.path.exists('./som_IP'):
-    shutil.rmtree('./som_IP_old', ignore_errors=True)
-    os.rename('./som_IP', './som_IP_old')
-
 if not args.skip_hls:
+
     # Synthesize HLS source codes
-    os.makedirs('./som_IP/Vivado_HLS')
+    shutil.rmtree('./Vivado_HLS', ignore_errors=True)
+    os.makedirs('./Vivado_HLS')
     used_resources = {True:{},False:{}}
 
-    msg.info('Synthesizing SmartOmpSsManager HLS sources')
+    msg.info('Synthesizing HLS sources')
     for file_ in glob.glob('./src/*.cpp'):
         synthesize_hls(file_, ['./src/som.hpp'])
 
     for file_ in glob.glob('./src/extended/*.cpp'):
         synthesize_hls(file_, ['./src/som.hpp'], True)
 
+if not args.skip_cutoff_gen:
+    if os.path.exists('./cutoff_IP'):
+        shutil.rmtree('./cutoff_IP_old', ignore_errors=True)
+        os.rename('./cutoff_IP', './cutoff_IP_old')
+
+    os.makedirs('./cutoff_IP/Vivado/CutoffManager')
+    os.makedirs('./cutoff_IP/IP_packager')
+
+    generate_cutoff_IP()
+
+if not args.skip_pom_gen:
+    if os.path.exists('./pom_IP'):
+        shutil.rmtree('./pom_IP_old', ignore_errors=True)
+        os.rename('./pom_IP', './pom_IP_old')
+
+    # Generate Vivado project and package IP
+    os.makedirs('./pom_IP/Vivado/PicosOmpSsManager')
+    os.makedirs('./pom_IP/IP_packager')
+
+    generate_POM_IP()
+
+if not args.skip_som_gen:
+    if os.path.exists('./som_IP'):
+        shutil.rmtree('./som_IP_old', ignore_errors=True)
+        os.rename('./som_IP', './som_IP_old')
+
     # Generate Vivado project and package IP
     os.makedirs('./som_IP/Vivado/SmartOmpSsManager')
     os.makedirs('./som_IP/IP_packager')
 
-    # SmartOmpSsManager utilization
-    f = open('./som_IP/IP_packager/som_resource_utilization.json', 'w')
-    f.write(json.dumps(used_resources[False]) + '\n')
+    if not args.skip_hls:
+        # SmartOmpSsManager utilization
+        f = open('./som_IP/IP_packager/som_resource_utilization.json', 'w')
+        f.write(json.dumps(used_resources[False]) + '\n')
+        f.close()
 
-    # Extended SmartOmpSsManager utilization
-    f = open('./som_IP/IP_packager/ext_som_resource_utilization.json', 'w')
-    f.write(json.dumps(used_resources[True]) + '\n')
+        # Extended SmartOmpSsManager utilization
+        f = open('./som_IP/IP_packager/ext_som_resource_utilization.json', 'w')
+        f.write(json.dumps(used_resources[True]) + '\n')
+        f.close()
 
-generate_IP()
+    generate_SOM_IP()
+
